+++ src/subjects/recent.jl
@@ -11,91 +11,67 @@ A variant of Subject that emits its recent value whenever it is subscribed to.
 
 See also: [`RecentSubjectFactory`](@ref), [`Subject`](@ref), [`SubjectFactory`](@ref)
 """
-function RecentSubject end
+mutable struct RecentSubject{D, S} <: Subscribable{D}
+    subject :: S
+    recent  :: Union{D, Blank}
+end
+
+getscheduler(subject::RecentSubject) = getscheduler(subject.subject)
+
+Base.show(io::IO, ::RecentSubject{D, S}) where { D, S } = print(io, "RecentSubject($D, $S)")
+
+Base.similar(subject::RecentSubject{D, S}) where { D, S } = RecentSubject(D, similar(subject.subject))
+
+RecentSubject(::Type{D}) where D = RecentSubject(D, SubjectFactory(AsapScheduler()))
+
+RecentSubject(::Type{D}, factory::F) where { D, F <: AbstractSubjectFactory } = RecentSubject(D, create_subject(D, factory))
+RecentSubject(::Type{D}, subject::S) where { D, S }                           = RecentSubject{D, S}(subject, blank)
+
+##
+
+getrecent(subject::RecentSubject)         = subject.recent
+setrecent!(subject::RecentSubject, value) = subject.recent = value
+
+##
+
+function on_next!(subject::RecentSubject{D}, data::D) where D
+    setrecent!(subject, data)
+    on_next!(subject.subject, data)
+end
+
+on_error!(subject::RecentSubject, err) = on_error!(subject.subject, err)
+on_complete!(subject::RecentSubject)   = on_complete!(subject.subject)
+
+##
+
+function on_subscribe!(subject::RecentSubject, actor)
+    recent = getrecent(subject)
+    if recent !== blank
+        on_next!(actor, recent)
+    end
+    return on_subscribe!(subject.subject, actor)
+end
+
+##
 
 """
     RecentSubjectFactory(factory::F) where { F <: AbstractSubjectFactory }
-    RecentSubjectFactory(; scheduler::H = AsapScheduler()) where { H <: AbstractScheduler }
+    RecentSubjectFactory(; scheduler::H = AsapScheduler()) where { H }
 
 A variant of SubjectFactory that creates an instance of RecentSubject.
 
 See also: [`SubjectFactory`](@ref), [`AbstractSubjectFactory`](@ref), [`RecentSubject`](@ref), [`Subject`](@ref)
 """
-function RecentSubjectFactory end
-
-##
-
-mutable struct RecentSubjectInstance{D, S} <: AbstractSubject{D}
-    subject :: S
-    recent  :: Union{D, Nothing}
-end
-
-Base.show(io::IO, ::RecentSubjectInstance{D, S}) where { D, S } = print(io, "RecentSubject($D, $S)")
-
-Base.similar(subject::RecentSubjectInstance{D, S}) where { D, S } = RecentSubject(D, similar(subject.subject))
-
-function RecentSubject(::Type{D}) where D
-    return RecentSubject(D, SubjectFactory(AsapScheduler()))
-end
-
-function RecentSubject(::Type{D}, factory::F) where { D, F <: AbstractSubjectFactory }
-    return RecentSubject(D, create_subject(D, factory))
-end
-
-function RecentSubject(::Type{D}, subject::S) where { D, S }
-    return as_recent_subject(D, as_subject(S), subject)
-end
-
-as_recent_subject(::Type{D},  ::InvalidSubjectTrait,    subject)    where D          = throw(InvalidSubjectTraitUsageError(subject))
-as_recent_subject(::Type{D1}, ::ValidSubjectTrait{D2},  subject)    where { D1, D2 } = throw(InconsistentSubjectDataTypesError{D1, D2}(subject))
-as_recent_subject(::Type{D},  ::ValidSubjectTrait{D},   subject::S) where { D, S }   = RecentSubjectInstance{D, S}(subject, nothing)
-
-##
-
-getrecent(subject::RecentSubjectInstance)         = subject.recent
-setrecent!(subject::RecentSubjectInstance, value) = subject.recent = value
-
-##
-
-function on_next!(subject::RecentSubjectInstance{D}, data::D) where D
-    setrecent!(subject, data)
-    next!(subject.subject, data)
-end
-
-function on_error!(subject::RecentSubjectInstance, err)
-    error!(subject.subject, err)
-end
-
-function on_complete!(subject::RecentSubjectInstance)
-    complete!(subject.subject)
-end
-
-##
-
-function on_subscribe!(subject::RecentSubjectInstance, actor)
-    recent = getrecent(subject)
-    if recent !== nothing
-        next!(actor, recent)
-    end
-    return subscribe!(subject.subject, actor)
-end
-
-##
-
-struct RecentSubjectFactoryInstance{ F <: AbstractSubjectFactory } <: AbstractSubjectFactory
+struct RecentSubjectFactory{ F <: AbstractSubjectFactory } <: AbstractSubjectFactory
     factory :: F
 end
 
-Base.show(io::IO, ::RecentSubjectFactoryInstance{F}) where F = print(io, "RecentSubjectFactory($F)")
+Base.show(io::IO, ::RecentSubjectFactory{F}) where F = print(io, "RecentSubjectFactory($F)")
 
-create_subject(::Type{L}, factory::RecentSubjectFactoryInstance) where L = RecentSubject(L, factory.factory)
+create_subject(::Type{L}, factory::RecentSubjectFactory) where L = RecentSubject(L, factory.factory)
 
-function RecentSubjectFactory(factory::F) where { F <: AbstractSubjectFactory }
-    return RecentSubjectFactoryInstance(factory)
-end
-
-function RecentSubjectFactory(; scheduler::H = AsapScheduler()) where { H <: AbstractScheduler }
-    return RecentSubjectFactoryInstance(SubjectFactory{H}(scheduler))
+function RecentSubjectFactory(; scheduler::H = AsapScheduler()) where { H }
+    return RecentSubjectFactory(SubjectFactory{H}(scheduler))
 end
 
 ##