+++ src/subjects/replay.jl
@@ -14,86 +14,70 @@ in addition to emitting new values to existing subscribers.
 
 See also: [`ReplaySubjectFactory`](@ref), [`Subject`](@ref), [`SubjectFactory`](@ref)
 """
-function ReplaySubject end
+struct ReplaySubject{D, S} <: Subscribable{D}
+    subject :: S
+    buffer  :: CircularBuffer{D}
+end
+
+getscheduler(subject::ReplaySubject) = getscheduler(subject.subject)
+
+Base.show(io::IO, ::ReplaySubject{D, S}) where { D, S } = print(io, "ReplaySubject($D, $S)")
+
+Base.similar(subject::ReplaySubject{D, S}) where { D, S } = ReplaySubject(D, capacity(subject.buffer), similar(subject.subject))
+
+ReplaySubject(::Type{D}, size::Int) where D = ReplaySubject(D, size, SubjectFactory(AsapScheduler()))
+
+ReplaySubject(::Type{D}, size::Int, factory::F) where { D, F <: AbstractSubjectFactory } = ReplaySubject(D, size, create_subject(D, factory))
+ReplaySubject(::Type{D}, size::Int, subject::S) where { D, S }                           = ReplaySubject{D, S}(subject, CircularBuffer{D}(size))
+
+##
+
+function on_next!(subject::ReplaySubject{D}, data::D) where D
+    if isactive(subject.subject)
+        push!(subject.buffer, data)
+        on_next!(subject.subject, data)
+    end
+end
+
+on_error!(subject::ReplaySubject, err) = on_error!(subject.subject, err)
+on_complete!(subject::ReplaySubject)   = on_complete!(subject.subject)
+
+##
+
+function on_subscribe!(subject::ReplaySubject, actor)
+    for value in subject.buffer
+        on_next!(actor, value)
+    end
+    return on_subscribe!(subject.subject, actor)
+end
+
+##
 
 """
     ReplaySubjectFactory(size::Int, factory::F) where { F <: AbstractSubjectFactory }
-    ReplaySubjectFactory(size::Int; scheduler::H = AsapScheduler()) where { H <: AbstractScheduler }
+    ReplaySubjectFactory(size::Int; scheduler::H = AsapScheduler()) where { H  }
 
 A variant of SubjectFactory that creates an instance of ReplaySubject.
 
 See also: [`SubjectFactory`](@ref), [`AbstractSubjectFactory`](@ref), [`ReplaySubject`](@ref), [`Subject`](@ref)
 """
-function ReplaySubjectFactory end
-
-struct ReplaySubjectInstance{D, S} <: AbstractSubject{D}
-    subject :: S
-    buffer  :: CircularBuffer{D}
-end
-
-Base.show(io::IO, ::ReplaySubjectInstance{D, S}) where { D, S } = print(io, "ReplaySubjectInstance($D, $S)")
-
-Base.similar(subject::ReplaySubjectInstance{D, S}) where { D, S } = ReplaySubject(D, capacity(subject.buffer), similar(subject.subject))
-
-function ReplaySubject(::Type{D}, size::Int) where D
-    return ReplaySubject(D, size, SubjectFactory(AsapScheduler()))
-end
-
-function ReplaySubject(::Type{D}, size::Int, factory::F) where { D, F <: AbstractSubjectFactory }
-    return ReplaySubject(D, size, create_subject(D, factory))
-end
-
-function ReplaySubject(::Type{D}, size::Int, subject::S) where { D, S }
-    return as_replay_subject(D, as_subject(S), size, subject)
-end
-
-as_replay_subject(::Type{D},  ::InvalidSubjectTrait,    size::Int, subject)    where D          = throw(InvalidSubjectTraitUsageError(subject))
-as_replay_subject(::Type{D1}, ::ValidSubjectTrait{D2},  size::Int, subject)    where { D1, D2 } = throw(InconsistentSubjectDataTypesError{D1, D2}(subject))
-as_replay_subject(::Type{D},  ::ValidSubjectTrait{D},   size::Int, subject::S) where { D, S }   = ReplaySubjectInstance{D, S}(subject, CircularBuffer{D}(size))
-
-##
-
-function on_next!(subject::ReplaySubjectInstance{D}, data::D) where D
-    if isactive(subject.subject)
-        push!(subject.buffer, data)
-        next!(subject.subject, data)
-    end
-end
-
-function on_error!(subject::ReplaySubjectInstance, err)
-    error!(subject.subject, err)
-end
-
-function on_complete!(subject::ReplaySubjectInstance)
-    complete!(subject.subject)
-end
-
-##
-
-function on_subscribe!(subject::ReplaySubjectInstance, actor)
-    for value in subject.buffer
-        next!(actor, value)
-    end
-    return subscribe!(subject.subject, actor)
-end
-
-##
-
-struct ReplaySubjectFactoryInstance{ F <: AbstractSubjectFactory } <: AbstractSubjectFactory
+struct ReplaySubjectFactory{ F <: AbstractSubjectFactory } <: AbstractSubjectFactory
     factory :: F
     size    :: Int
 end
 
-Base.show(io::IO, subject::ReplaySubjectFactoryInstance{F}) where F = print(io, "ReplaySubjectFactoryInstance($F, size = $(subject.size))")
 
-create_subject(::Type{L}, factory::ReplaySubjectFactoryInstance) where L = ReplaySubject(L, factory.size, factory.factory)
+
+Base.show(io::IO, subject::ReplaySubjectFactory{F}) where F = print(io, "ReplaySubjectFactory($F, size = $(subject.size))")
+
+create_subject(::Type{L}, factory::ReplaySubjectFactory) where L = ReplaySubject(L, factory.size, factory.factory)
 
 function ReplaySubjectFactory(size::Int, factory::F) where { F <: AbstractSubjectFactory }
-    return ReplaySubjectFactoryInstance(factory, size)
+    return ReplaySubjectFactory{F}(factory, size)
 end
 
-function ReplaySubjectFactory(size::Int; scheduler::H = AsapScheduler()) where { H <: AbstractScheduler }
-    return ReplaySubjectFactoryInstance(SubjectFactory{H}(scheduler), size)
+function ReplaySubjectFactory(size::Int; scheduler::H = AsapScheduler()) where { H }
+    return ReplaySubjectFactory(size, SubjectFactory{H}(scheduler))
 end
 
 ##