+++ test/subjects/test_subject_pending.jl
@@ -60,7 +60,7 @@ using Rocket
         actor2 = keep(Int)
 
         values = Int[]
-        source = from(1:5) |> tap(d -> push!(values, d))
+        source = from_iterable(1:5) |> tap(d -> push!(values, d))
 
         subscription1 = subscribe!(subject, actor1)
         subscription2 = subscribe!(subject, actor2)
@@ -116,7 +116,7 @@ using Rocket
         actor2 = keep(Int)
 
         values = Int[]
-        source = from(1:5) |> tap(d -> push!(values, d))
+        source = from_iterable(1:5) |> tap(d -> push!(values, d))
 
         subscription1 = subscribe!(subject, actor1)
         subscription2 = subscribe!(subject, actor2)
@@ -136,7 +136,7 @@ using Rocket
         subject2 = similar(subject1)
 
         @test subject1 !== subject2
-        @test typeof(subject2) <: Rocket.PendingSubjectInstance
+        @test typeof(subject2) <: Rocket.PendingSubject
         @test eltype(subject2) === Int
 
         actor1 = keep(Int)