+++ test/subjects/test_subject_recent.jl
@@ -41,7 +41,7 @@ using Rocket
         actor2 = keep(Int)
 
         values = Int[]
-        source = from(1:5) |> tap(d -> push!(values, d))
+        source = from_iterable(1:5) |> tap(d -> push!(values, d))
 
         subscription1 = subscribe!(subject, actor1)
         subscription2 = subscribe!(subject, actor2)
@@ -97,7 +97,7 @@ using Rocket
         actor2 = keep(Int)
 
         values = Int[]
-        source = from(1:5) |> tap(d -> push!(values, d))
+        source = from_iterable(1:5) |> tap(d -> push!(values, d))
 
         subscription1 = subscribe!(subject, actor1)
         subscription2 = subscribe!(subject, actor2)
@@ -117,7 +117,7 @@ using Rocket
         subject2 = similar(subject1)
 
         @test subject1 !== subject2
-        @test typeof(subject2) <: Rocket.RecentSubjectInstance
+        @test typeof(subject2) <: Rocket.RecentSubject
         @test eltype(subject2) === Int
 
         actor1 = keep(Int)
@@ -147,7 +147,7 @@ using Rocket
         subject3 = similar(subject2)
 
         @test subject2 !== subject3
-        @test typeof(subject3) <: Rocket.RecentSubjectInstance
+        @test typeof(subject3) <: Rocket.RecentSubject
         @test eltype(subject3) === Int
 
         actor3 = keep(Int)