+++ test/subjects/test_subject_replay.jl
@@ -52,7 +52,7 @@ using Rocket
         actor2 = keep(Int)
 
         values = Int[]
-        source = from(1:5) |> tap(d -> push!(values, d))
+        source = from_iterable(1:5) |> tap(d -> push!(values, d))
 
         subscription1 = subscribe!(subject, actor1)
         subscription2 = subscribe!(subject, actor2)
@@ -115,7 +115,7 @@ using Rocket
         actor2 = keep(Int)
 
         values = Int[]
-        source = from(1:5) |> tap(d -> push!(values, d))
+        source = from_iterable(1:5) |> tap(d -> push!(values, d))
 
         subscription1 = subscribe!(subject, actor1)
         subscription2 = subscribe!(subject, actor2)
@@ -141,7 +141,7 @@ using Rocket
         subject2 = similar(subject1)
 
         @test subject1 !== subject2
-        @test typeof(subject2) <: Rocket.ReplaySubjectInstance
+        @test typeof(subject2) <: Rocket.ReplaySubject
         @test eltype(subject2) === Int
 
         actor1 = keep(Int)
@@ -173,7 +173,7 @@ using Rocket
         subject3 = similar(subject2)
 
         @test subject2 !== subject3
-        @test typeof(subject3) <: Rocket.ReplaySubjectInstance
+        @test typeof(subject3) <: Rocket.ReplaySubject
         @test eltype(subject3) === Int
 
         actor3 = keep(Int)